% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{methods}
\name{counts,SNPhood-method}
\alias{counts}
\alias{counts,SNPhood-method}
\title{Count method for SNPhood}
\usage{
\S4method{counts}{SNPhood}(object, type, readGroup = NULL, dataset = NULL,
  ...)
}
\arguments{
\item{object}{object}

\item{type}{type}

\item{readGroup}{readGroup}

\item{dataset}{dataset}

\item{...}{extra params}
}
\description{
Count method for SNPhood
}

